apps/api/requirements.txt
# fastapi/uvicorn versions align with starlette>=0.46 (required by browser-use dependency tree)
fastapi==0.135.3
uvicorn[standard]==0.43.0
sqlmodel==0.0.22
httpx==0.28.1
python-dotenv>=1.0.1
pydantic>=2.9.0,<3
playwright==1.49.1
browser-use==0.12.6
